+++ src/Aws/Iam/Exception/EntityTemporarilyUnmodifiableException.php
@@ -17,8 +17,6 @@
 namespace Aws\Iam\Exception;
 
 /**
- * The request was rejected because it referenced an entity that is temporarily unmodifiable, such as a user name that
- * was deleted and then recreated. The error indicates that the request is likely to succeed if you try again after wai
- * ting several minutes. The error message describes the entity.
+ * The request was rejected because it referenced an entity that is temporarily unmodifiable, such as a user name that was deleted and then recreated. The error indicates that the request is likely to succeed if you try again after waiting several minutes. The error message describes the entity.
  */
 class EntityTemporarilyUnmodifiableException extends IamException {}