+++ src/Aws/Iam/Exception/InvalidAuthenticationCodeException.php
@@ -17,7 +17,6 @@
 namespace Aws\Iam\Exception;
 
 /**
- * The request was rejected because the authentication code was not recognized. The error message describes the specifi
- * c error.
+ * The request was rejected because the authentication code was not recognized. The error message describes the specific error.
  */
 class InvalidAuthenticationCodeException extends IamException {}