+++ src/Aws/Iam/Exception/MalformedCertificateException.php
@@ -17,7 +17,6 @@
 namespace Aws\Iam\Exception;
 
 /**
- * The request was rejected because the certificate was malformed or expired. The error message describes the specific
- * error.
+ * The request was rejected because the certificate was malformed or expired. The error message describes the specific error.
  */
 class MalformedCertificateException extends IamException {}